requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
pandas>=2.0.0
python-dotenv
openpyxl